Reverted rename of _azure
