+++ azext_iot/common/_azure.py
