+++ azext_iot/operations/hub.py
@@ -21,7 +21,7 @@ from azext_iot.common.sas_token_auth import SasTokenAuthentication
 from azext_iot.common.shared import (DeviceAuthType,
                                      SdkType,
                                      MetricType)
-from azext_iot.common.azure import get_iot_hub_connection_string
+from azext_iot.common._azure import get_iot_hub_connection_string
 from azext_iot.common.utility import (shell_safe_json_parse,
                                       validate_key_value_pairs, url_encode_dict,
                                       evaluate_literal, unpack_msrest_error,
@@ -907,7 +907,7 @@ def iot_get_sas_token(cmd, hub_name=None, device_id=None, policy_name='iothubown
 
 def _iot_build_sas_token(cmd, hub_name=None, device_id=None, module_id=None, policy_name='iothubowner',
                          key_type='primary', duration=3600, resource_group_name=None, login=None):
-    from azext_iot.common.azure import (parse_iot_device_connection_string,
+    from azext_iot.common._azure import (parse_iot_device_connection_string,
                                         parse_iot_device_module_connection_string)
 
     target = get_iot_hub_connection_string(cmd, hub_name, resource_group_name, policy_name, login=login)
